+++ PerksCSharp/PerkProject/Plugin.cs
@@ -174,6 +174,7 @@ namespace PerkManager
             medsTexts[perkStem + "taunt1e"] = "Taunt on this hero can stack and increases damage by 1 per charge.";
             medsTexts[perkStem + "taunt1f"] = "Taunt on heroes increases maximum Powerful by 1 per charge.";
             medsTexts[perkStem + "taunt1g"] = "Taunt on you is not lost at the start of turn, but 1 charge is consumed when you take damage.";
+            medsTexts[perkStem + "taunt1h"] = "Taunt on monsters decreases All Resistances by 5% per charge.";
 
             // Custom Text for Fortify
 
@@ -199,6 +200,7 @@ namespace PerkManager
 
             // Custom Text for Burn
             medsTexts[perkStem + "burn2e"] = "Burn increases the damage dealt by Dark explosions by 0.5% per charge.";
+            medsTexts[perkStem + "burn2f"] = "When Burn is Dispelled from a Monster, they suffer Fire damage to the target equal to twice the Burn that was purged. (Needs testing)";
             medsTexts["burn2d"] = "Burn deals double damage on enemies with 3 or less curses (burn included).";
 
 
@@ -344,6 +346,7 @@ namespace PerkManager
             medsTexts[perkStem + "leech0d"] = "Charges applied +1. Decrease healing done by Leech by 50%.";
             medsTexts[perkStem + "leech0e"] = "Leech explodes at the end of turn.";
             medsTexts[perkStem + "leech0f"] = "Increase curses applied by Leech by 100%. Leech no longer reduces enemy resistances.";
+            medsTexts[perkStem + "leech0g"] = "Rather than healing, when Leech explodes, all enemies lose HP equal to the target's Bleed.";
 
 
             // apply patches